Fix bug in Inventory sidebar heading, total item count and weight.
Apparently there are "fake" non-item entries in the list of items
that the sidebar has access to, which have a weight of -1, which
is some kind of sentinel value that means "this isn't actually an
item that the player is really carrying".  So skip those ones.

+++ libnethack/src/cmd.c
@@ -871,7 +871,7 @@ doattributes(const struct nh_cmd_arg *arg)
     add_menutext(&menu, buf);
 
     wc = weight_cap();
-    buf = msgprintf("%-10s: %ld (", "burden", wc + inv_weight());
+    buf = msgprintf("%-10s: %ld (", "burden", wc + (long) inv_weight());
 
     switch (calc_capacity(wc / 4)) {
     case UNENCUMBERED:
